pyproject.toml
[project]
name = "matharena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mislav Balunovic", email = "mislav.balunovic@gmail.com" },
    { name = "Jasper Dekoninck", email = "jasper.dekoninck@inf.ethz.ch" },
    { name = "Nikola Jovanovic", email = "nikola.jovanovic@inf.ethz.ch" },
    { name = "Ivo Petrov", email = "ivo.petrov@insait.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "beautifulsoup4>=4.13.1",
    "google-genai>=1.11.0",
    "loguru>=0.7.3",
    "openai>=1.102.0",
    "python-fasthtml>=0.12.1",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "together>=1.3.14",
    "antlr4-python3-runtime==4.11",
    "sympy >=1.13.1",
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
    "matplotlib>=3.9.3",
    "json5>=0.10.0",
    "datasets >= 3.5.0",
    "modal>=1.0.0",
    "pytest>=8.3.5",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.1",
    "thefuzz>=0.22.1",
    "zstandard>=0.23.0",
    "sentence_transformers>=5.0.0",
    "transformers>=4.55.0",
    "pyyaml>=6.0",
    "openai-harmony>=0.0.4",
    "pyarrow>=20.0.0",
    "docker>=7.1.0",
    "vllm>=0.13.0",
    "PyMuPDF>=1.26.7"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["website"]

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "E266", "E501", "W503"]
